+++ src/Home/Sidebar/Sidebar.js
@@ -12,11 +12,7 @@ import ListItemIcon from '@material-ui/core/ListItemIcon';
 import ListItemText from '@material-ui/core/ListItemText';
 import InboxIcon from '@material-ui/icons/MoveToInbox';
 import MailIcon from '@material-ui/icons/Mail';
-import InputLabel from '@material-ui/core/InputLabel';
-import MenuItem from '@material-ui/core/MenuItem';
-import FormControl from '@material-ui/core/FormControl';
-import Select from '@material-ui/core/Select';
-import Button from '@material-ui/core/Button';
+import Bench from "./SidebarComponents/Bench";
 
 const drawerWidth = 240;
 
@@ -32,62 +28,14 @@ const useStyles = makeStyles(theme => ({
   drawerPaper: {
     width: drawerWidth,
   },
-  button: {
-    display: 'block',
-    marginTop: theme.spacing(2),
-  },
-  formControl: {
-    margin: theme.spacing(1),
-    minWidth: 120,
-  },
+
 }));
 
 export default function PermanentDrawerLeft() {
   const classes = useStyles();
 
-  const [age, setAge] = React.useState('');
-  const [open, setOpen] = React.useState(false);
-
-  function handleChange(event) {
-    setAge(event.target.value);
-  }
-
-  function handleClose() {
-    setOpen(false);
-  }
-
-  function handleOpen() {
-    setOpen(true);
-  }
   return (
-      
     <div className={classes.root}>
-        <form autoComplete="off">
-      <Button className={classes.button} onClick={handleOpen}>
-        Open the select
-      </Button>
-      <FormControl className={classes.formControl}>
-        <InputLabel htmlFor="demo-controlled-open-select">Age</InputLabel>
-        <Select
-          open={open}
-          onClose={handleClose}
-          onOpen={handleOpen}
-          value={age}
-          onChange={handleChange}
-          inputProps={{
-            name: 'age',
-            id: 'demo-controlled-open-select',
-          }}
-        >
-          <MenuItem value="">
-            <em>None</em>
-          </MenuItem>
-          <MenuItem value={10}>Ten</MenuItem>
-          <MenuItem value={20}>Twenty</MenuItem>
-          <MenuItem value={30}>Thirty</MenuItem>
-        </Select>
-      </FormControl>
-    </form>
       <CssBaseline />
       {/* <AppBar position="fixed" className={classes.appBar}>
         <Toolbar>
@@ -105,27 +53,38 @@ export default function PermanentDrawerLeft() {
         anchor="left"
       >
         <div className={classes.toolbar} />
+        <Divider />
+        <List>
+          {['Offense'].map((text, index) => (
+            <ListItem button key={text}>
+              <ListItemIcon>{index % 1 === 0 ? <InboxIcon /> : <MailIcon />}</ListItemIcon>
+              <ListItemText primary={text} />
+            </ListItem>
+          ))}
+          <Bench/>
+        </List>
         
         <Divider />
         <List>
-            
-          {['Offense', 'Defrense', 'Kicker'].map((text, index) => (
+          {["Defense"].map((text, index) => (
             <ListItem button key={text}>
-                
               <ListItemIcon>{index % 2 === 0 ? <InboxIcon /> : <MailIcon />}</ListItemIcon>
               <ListItemText primary={text} />
             </ListItem>
           ))}
+          <Bench/>
         </List>
-        <Divider />
         <List>
-          {['All mail', 'Trash', 'Spam'].map((text, index) => (
+            <Divider/>
+          {["Kicker"].map((text, index) => (
             <ListItem button key={text}>
               <ListItemIcon>{index % 2 === 0 ? <InboxIcon /> : <MailIcon />}</ListItemIcon>
               <ListItemText primary={text} />
             </ListItem>
           ))}
+          <Bench/>
         </List>
+        <Divider/>
       </Drawer>
 
     </div>
